Fix incorrect batching

+++ common/src/main/java/cn/zbx1425/mtrsteamloco/sound/BveTrainSoundFix.java
@@ -65,7 +65,7 @@ public class BveTrainSoundFix extends TrainSoundBase {
         }
 
         // Get door delay of the first sec off
-        final int dwellTicks = totalDwellTicks - 20;
+        final int dwellTicks = (int) (totalDwellTicks - 20);
         final float stopTicks = elapsedDwellTicks - 20;
 
         final SoundEvent soundEvent;
